move concentric localization into new file
