+++ nbed/localizers/__init__.py
@@ -1,15 +1,17 @@
 """Init Localizer classes."""
 
 from .ace import ACELocalizer
-from .base import Localizer
-from .pyscf import BOYSLocalizer, IBOLocalizer, PMLocalizer
-from .spade import SPADELocalizer
+from .occupied.base import OccupiedLocalizer
+from .occupied.pyscf import BOYSLocalizer, IBOLocalizer, PMLocalizer
+from .occupied.spade import SPADELocalizer
+from .virtual.concentric import ConcentricLocalizer
 
 __all__ = [
-    "Localizer",
     "BOYSLocalizer",
     "IBOLocalizer",
     "PMLocalizer",
     "SPADELocalizer",
     "ACELocalizer",
+    "ConcentricLocalizer",
+    "OccupiedLocalizer",
 ]
